requirements.txt
pymongo==3.4.0
tornado==4.4.2
motor==1.1
mapnik==0.1
numpy==1.11.3
geojson==1.3.3
PyYAML==3.12
redis==2.10.5
pyTelegramBotAPI==3.2.0
scrypt==0.8.0
